chore: primer refactor
